# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parvath FinServ CRM",
    "version": "1.0.0",
    "description": "All business endpoints require a workspace-scoped session. Mutations require the CSRF token returned by /auth/csrf or /auth/login. Dates use Asia/Kolkata; money is an integer string in paise. See docs/api.md for lifecycle contracts."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "parvath.sid"
      }
    },
...
  },
  "paths": {
...
    "/auth/csrf": {
      "get": {
        "summary": "Establish CSRF token and anonymous session",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Destroy session",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Current membership",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/account": {
      "patch": {
        "summary": "Update name and optional password; currentPassword required",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Send reset link through configured SMTP",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Consume single-use reset token",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 64,
                    "maxLength": 64
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12
                  }
                },
                "required": [
                  "token",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
  }
}